Bidirectional_WordPress_Sync/requirements.txt
# Bidirectional WordPress ↔ GitHub Pages Sync
# Python dependencies

# Core dependencies
pyyaml>=6.0
requests>=2.31.0
python-frontmatter>=1.0.0
markdown>=3.4.0
gitpython>=3.1.0

# HTML/Markdown processing
beautifulsoup4>=4.12.0
html2text>=2020.1.16
markdownify>=0.11.0

# Date/time handling
python-dateutil>=2.8.0

# Logging and utilities
colorama>=0.4.6
tqdm>=4.65.0

# Optional: For notifications
# requests[security]  # For webhook notifications
# python-dotenv>=1.0.0  # For environment variables